# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI will API
  description: |
    AI will - AI音声会話アプリのREST API仕様書

    ## 認証
    JWT Bearer Tokenを使用します。
    ```
    Authorization: Bearer {access_token}
    ```

    ## エラーレスポンス
    すべてのエラーは統一フォーマットで返却されます。
    ```json
    {
      "error": {
        "code": "error_code",
        "message": "エラーメッセージ",
        "details": []
      }
    }
    ```

    ## ページング
    Cursor方式を採用。`cursor`と`limit`パラメータを使用。
    ```json
    {
      "data": [...],
      "pagination": {
        "next_cursor": "eyJpZCI6...",
        "has_more": true
      }
    }
    ```

    ## 冪等性（Idempotency）
    購入系・ジョブ作成系APIでは`Idempotency-Key`ヘッダーが必須です。
    - キー: クライアント生成のUUID
    - 有効期限: 24時間
    - 同一キーで再リクエスト時は同一レスポンスを返却

    ## ストリーミング（SSE）
    会話APIでは Server-Sent Events によるストリーミングをサポートします。
    - エンドポイント: `POST /threads/{id}/messages:stream`
    - イベント: `message_start`, `content_delta`, `message_done`, `error`

    ## 主要なユースケース例
    - **Example 1**: 会話メッセージ送信（`POST /threads/{id}/messages`）
    - **Example 2**: 購入反映・Idempotency-Key付き（`POST /purchases`）
    - **Example 3**: クリップ作成（`POST /clips`）
  version: 1.3.0
  contact:
    name: AI will Support
    email: support@aiwill.app
  license:
    name: Proprietary
    url: https://aiwill.app/terms
servers:
  - url: https://api.aiwill.app/v1
    description: Production
  - url: https://api.staging.aiwill.app/v1
    description: Staging
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: ユーザー登録
      description: メールアドレスとパスワードで新規ユーザーを登録します。
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: user@example.com
              password: SecureP@ss123
      responses:
        '201':
          description: 登録成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /auth/login:
    post:
      tags: [Auth]
      summary: ログイン
      description: メールアドレスとパスワードでログインします。
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecureP@ss123
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: 認証失敗
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: invalid_credentials
                  message: メールアドレスまたはパスワードが正しくありません
                  details: []
        '423':
          description: アカウントロック
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: トークン更新
      description: リフレッシュトークンを使用してアクセストークンを更新します。
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: トークン無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Auth]
      summary: ログアウト
      description: リフレッシュトークンを無効化してログアウトします。
      operationId: logout
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: ログアウト成功
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me:
    get:
      tags: [Auth]
      summary: ユーザー情報取得
      description: 認証ユーザーの情報とオンボーディング状態を取得します。
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags: [Auth]
      summary: プロフィール更新
      description: ユーザーのプロフィール情報を更新します。
      operationId: updateMe
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me/consent:
    post:
      tags: [Auth]
      summary: 利用規約同意
      description: 利用規約とプライバシーポリシーへの同意を記録します。
      operationId: consent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentRequest'
      responses:
        '200':
          description: 同意完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /me/age-verify:
    post:
      tags: [Auth]
      summary: 年齢確認
      description: ユーザーの年齢を確認し、適切な制限を設定します。
      operationId: ageVerify
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgeVerifyRequest'
      responses:
        '200':
          description: 年齢確認完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgeVerifyResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: 同意未完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          $ref: '#/components/responses/ConflictError'

  # ----------------------------------------------------------
  # Catalog
  # ----------------------------------------------------------
  /packs:
    get:
      tags: [Catalog]
      summary: Pack一覧取得
      description: Persona/Scenario Packの一覧を取得します。フィルタ・ソート・ページング対応。
      operationId: listPacks
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [persona, scenario]
          description: Pack種別
        - name: query
          in: query
          schema:
            type: string
          description: キーワード検索
        - name: tags
          in: query
          schema:
            type: string
          description: タグ絞り込み（カンマ区切り）
        - name: age_rating
          in: query
          schema:
            type: string
            enum: [all, r15, r18]
          description: |
            年齢レーティングフィルタ。
            ユーザーの age_group に応じて閲覧可能な Pack が自動フィルタされる:
            - u13: all のみ
            - u18: all, r15
            - adult: すべて
        - $ref: '#/components/parameters/sort'
        - $ref: '#/components/parameters/order'
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackListResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /packs/{pack_id}:
    get:
      tags: [Catalog]
      summary: Pack詳細取得
      description: 指定したPackの詳細情報を取得します。
      operationId: getPack
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/packId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/AgeRestrictedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /packs/{pack_id}/items:
    get:
      tags: [Catalog]
      summary: Pack構成アイテム取得
      description: Packに含まれるキャラクター、イベント等の構成アイテムを取得します。
      operationId: getPackItems
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/packId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackItemsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /tags:
    get:
      tags: [Catalog]
      summary: タグ一覧取得
      description: 利用可能なタグの一覧を取得します。
      operationId: listTags
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [pack, character]
          description: タグ種別
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ----------------------------------------------------------
  # Conversation
  # ----------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token

  # ----------------------------------------------------------
  # Parameters
  # ----------------------------------------------------------
  parameters:
    packId:
      name: pack_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Pack ID
...
    cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: ページングカーソル
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 取得件数
    sort:
      name: sort
      in: query
      schema:
        type: string
        default: created_at
      description: ソートキー
    order:
      name: order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
      description: 昇順/降順
...
  responses:
    ValidationError:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: validation_error
              message: リクエストの検証に失敗しました
              details:
                - field: email
                  code: invalid_format
                  message: 有効なメールアドレス形式で入力してください
    UnauthorizedError:
      description: 認証エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: unauthenticated
              message: 認証が必要です
              details: []
    ForbiddenError:
      description: 権限エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: forbidden
              message: この操作を実行する権限がありません
              details: []
    AgeRestrictedError:
      description: 年齢制限エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: age_restricted
              message: このコンテンツは年齢制限があります
              details: []
...
    NotFoundError:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: not_found
              message: リソースが見つかりません
              details: []
    ConflictError:
      description: 競合エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: conflict
              message: リソースが既に存在します
              details: []
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: エラーコード
            message:
              type: string
              description: エラーメッセージ
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  code:
                    type: string
                  message:
                    type: string
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
          description: 次ページのカーソル
        has_more:
          type: boolean
          description: 次ページの有無

    # Auth
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
    Tokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: 有効期限（秒）
    TokenResponse:
      type: object
      properties:
        tokens:
          $ref: '#/components/schemas/Tokens'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        display_name:
          type: string
          nullable: true
        consent_at:
          type: string
          format: date-time
          nullable: true
        age_verified_at:
          type: string
          format: date-time
          nullable: true
        age_group:
          type: string
          enum: [u13, u18, adult]
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Onboarding:
      type: object
      properties:
        consent_completed:
          type: boolean
        age_verified:
          type: boolean
        completed:
          type: boolean
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/Tokens'
    UserResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
    MeResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        onboarding:
          $ref: '#/components/schemas/Onboarding'
    UpdateMeRequest:
      type: object
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 50
    ConsentRequest:
      type: object
      required: [terms_version, privacy_version]
      properties:
        terms_version:
          type: string
        privacy_version:
          type: string
    ConsentResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            consent_at:
              type: string
              format: date-time
        onboarding:
          $ref: '#/components/schemas/Onboarding'
    AgeVerifyRequest:
      type: object
      properties:
        birth_date:
          type: string
          format: date
          description: 生年月日（YYYY-MM-DD）
        age_group:
          type: string
          enum: [u13, u18, adult]
          description: 年齢区分（birth_dateの代替）
    Restrictions:
      type: object
      properties:
        adult_content:
          type: boolean
        purchase_limit:
          type: integer
          nullable: true
          description: 課金上限（円/月）
    AgeVerifyResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            age_verified_at:
              type: string
              format: date-time
            age_group:
              type: string
        onboarding:
          $ref: '#/components/schemas/Onboarding'
        restrictions:
          $ref: '#/components/schemas/Restrictions'

    # Catalog
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        count:
          type: integer
    Creator:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
          nullable: true
    PackStats:
      type: object
      properties:
        favorite_count:
          type: integer
        conversation_count:
          type: integer
        review_count:
          type: integer
        average_rating:
          type: number
          format: float
    Pack:
      type: object
      properties:
        id:
          type: string
          format: uuid
        pack_type:
          type: string
          enum: [persona, scenario]
        name:
          type: string
        description:
          type: string
        thumbnail_url:
          type: string
          format: uri
        cover_url:
          type: string
          format: uri
          nullable: true
        sample_voice_url:
          type: string
          format: uri
          nullable: true
        price:
          type: integer
          description: 価格（円）
        is_free:
          type: boolean
        age_rating:
          type: string
          enum: [all, r15, r18]
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        creator:
          $ref: '#/components/schemas/Creator'
        stats:
          $ref: '#/components/schemas/PackStats'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserStatus:
      type: object
      properties:
        owned:
          type: boolean
        favorited:
          type: boolean
    PackListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Pack'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PackDetailResponse:
      type: object
      properties:
        pack:
          $ref: '#/components/schemas/Pack'
        user_status:
          $ref: '#/components/schemas/UserStatus'
    PackItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        item_type:
          type: string
          enum: [character, event, voice_pack]
        item_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
    PackItemsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PackItem'
    TagListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Tag'

    # Conversation
...